stagehand PR review changes
